changed IntSliderControl.java and FloatSliderControl.java to sodium's SliderControl to fix some options being static with reese's sodium options

+++ src/main/java/ca/fxco/moreculling/config/MoreCullingConfig.java
@@ -33,7 +33,7 @@ public class MoreCullingConfig implements ConfigData {
 
     public boolean useItemFrame3FaceCulling = true;
 
-    public float itemFrame3FaceCullingRange = 4F; // default 4 blocks away
+    public int itemFrame3FaceCullingRange = 4; // default 4 blocks away
 
     public LeavesCullingMode leavesCullingMode = LeavesCullingMode.DEFAULT;
